+++ srcs/vector_dot.c
@@ -6,7 +6,7 @@
 /*   By: ddu-toit <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/07/04 08:19:55 by ddu-toit          #+#    #+#             */
-/*   Updated: 2016/07/11 13:50:15 by ddu-toit         ###   ########.fr       */
+/*   Updated: 2016/08/15 09:59:05 by ddu-toit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 ** Multiply two vectors and return the resulting scalar (dot product)
 */
 
-float	vector_dot(t_vector *v1, t_vector *v2)
+double	vector_dot(t_vector *v1, t_vector *v2)
 {
 	return (v1->x * v2->x + v1->y * v2->y + v1->z * v2->z);
 }
